pyproject.toml
[project]
name = "sparsecity"
version = "0.1.0"
description = "Sparse model training with torch compile and heavyball optimizers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.2.0",
    "heavyball>=1.5.2",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "jaxtyping>=0.2.36",
    "schedulefree>=1.4",
    "sentence-transformers>=3.4.0",
    "torch>=2.5.1",
    "tqdm>=4.67.1",
    "transformers>=4.48.0",
    "wandb>=0.19.4",
]
